template/requirements.txt
Flask
python-dotenv
yapf
ruff
